+++ tests/testthat/test-verb-filter.R
@@ -34,6 +34,13 @@ test_that("only add step if necessary", {
   expect_equal(lf %>% filter(), lf)
 })
 
+test_that("errors for named input", {
+  lf <- lazy_frame(x = 1, y = 2)
+
+  expect_snapshot(error = TRUE, filter(lf, x = 1))
+  expect_snapshot(error = TRUE, filter(lf, y > 1, x = 1))
+})
+
 test_that(".preserve is not supported", {
   lf <- lazy_frame(x = 1:3, y = 1:3)
   expect_snapshot(error = TRUE, lf %>% filter(x == 1, .preserve = TRUE))